requirements.txt
streamlit==1.38.0
# pandas==2.2.2
# numpy==1.26.4
python-docx==1.1.2
mammoth==1.8.0
openpyxl==3.1.2